fixed pull request changes
