+++ admin_showing.html
@@ -8,8 +8,7 @@
 <body>
 <div class="container">
     <div class="container-showing">
-       <!-- <form id="showing_form" action="https://kinoxp-back.azurewebsites.net/showings" method="post"> -->
-        <form id="showing_form" action="http://localhost:8080/showings" method="post">
+        <form id="showing_form" action="https://kinoxp-back.azurewebsites.net/showings" method="post">
             <input id="id" name="showingID" type="hidden">
 
             <label for="showing_date">Date</label>
