+++ js/module.js
@@ -1,17 +1,33 @@
 const ageLimits = ["A", "PG7", "PG11", "PG15", "F"];
 const movieCategories = ["ACTION", "COMEDY", "DRAMA", "ROMANCE", "HORROR", "SCIFI", "FANTASY", "SUPERHERO"];
 
-function fillDropdown(dropdownId, array, defaultText) {
-    const firstElement = document.createElement('option');
+function setTextContent(firstElement, defaultText) {
     firstElement.textContent = `Select ${defaultText}`;
     firstElement.disabled = true;
     firstElement.selected = true;
+}
+
+function fillDropdown(dropdownId, array, defaultText) {
+    const firstElement = document.createElement('option');
+    setTextContent(firstElement, defaultText)
     dropdownId.appendChild(firstElement);
 
     array.forEach(element => {
         const optionElement = document.createElement('option');
         optionElement.textContent = element;
         optionElement.value = element;
+        dropdownId.appendChild(optionElement);
+    });
+}
+
+function fillObjectDropdown(dropdownId, array, defaultText, txtContent, elementValue) {
+    const firstElement = document.createElement('option');
+    setTextContent(firstElement, defaultText)
+    dropdownId.appendChild(firstElement);
+    array.forEach(element => {
+        const optionElement = document.createElement('option');
+        optionElement.textContent = element[txtContent];
+        optionElement.value = element[elementValue];
         dropdownId.appendChild(optionElement);
     });
 }
@@ -23,10 +39,7 @@ async function postOrPutObjectAsJson(url, object, HttpVerb) {
         headers: {'Content-type' : 'application/json'},
         body: objectToJsonString
     }
-
-
     const response = await fetch(url, fetchOption);
-
     return response;
 }
 
@@ -40,4 +53,4 @@ async function restDelete(url) {
 }
 
 
-export {fillDropdown, postOrPutObjectAsJson, restDelete, ageLimits, movieCategories}
+export {fillDropdown, fillObjectDropdown, postOrPutObjectAsJson, restDelete, ageLimits, movieCategories}
